test_api/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
rules==2.2
furl==2.1
django-filter==2.4
